add resultManage
