+++ TeSystemProject/src/components/taskManage/verifyPerform/taskBaseInfo.vue
@@ -193,9 +193,9 @@
                                                            :on-change="handleChangeInter"
                                                            :on-error="uploadOnError"
                                                            accept=".xml"
-                                                           on-success="successUpload"
+                                                         
                                                            multiple
-                                                            :limit="1"
+                                                            :limit="1" 
                                                             :on-exceed="handleExceed"
                                                             :file-list="fileList"
                                                             :show-file-list="true"
@@ -415,6 +415,7 @@ import logo from '@/components/logo'
 import testCaseNavBar from '@/components/taskManage/testCaseNavBar'
 import leftBar from '@/components/leftBar'
 import {Message} from 'element-ui'
+
 export default 
   {
      name: "ProjectList",
@@ -431,13 +432,8 @@ export default
                 registerFile:''
               } ,
               characteristic:[]  
-            },
-          
-       // servicesInfos:[],
-
-       
+            },     
           servicesInfos: [],//服务命名特征
-
           datasInfos: [],//服务数据列表
          activeNames: ['1'],//折叠框默认展开1
 
@@ -453,7 +449,7 @@ export default
         loading_bottom: null,
         pagesize: 5,
         curPage: 1,
-        search: '',
+       // search: '',
         // 项目对应的多个任务的对话框，内置了表格
         dialogTableVisible: false,
        
@@ -488,11 +484,11 @@ export default
       },
     computed: //赋值
     {
-      global_url() {
-        return this.$store.state.global_url
+     /* global_url() {
+      return this.$store.state.global_url
       },
       // 模糊搜索
-      fuzzy() {
+   /*   fuzzy() {
         const search = this.search;
         if (search) {
           // filter() 方法创建一个新的数组，
@@ -515,32 +511,9 @@ export default
           })
         }
         return this.servicesInfos;
-      },
+      },*/
 
-       fuzzyd() {
-        const search = this.search;
-        if (search) {
-          // filter() 方法创建一个新的数组，
-          // 新数组中的元素是通过检查指定数组中符合条件的所有元素。
-          // 注意： filter() 不会对空数组进行检测。
-          // 注意： filter() 不会改变原始数组。
-        
-          return this.datasInfos.filter(data => {
-            // some() 方法用于检测数组中的元素是否满足指定条件;
-            // some() 方法会依次执行数组的每个元素：
-            // 如果有一个元素满足条件，则表达式返回true , 剩余的元素不会再执行检测;
-            // 如果没有满足条件的元素，则返回false。
-            // 注意： some() 不会对空数组进行检测。
-            // 注意： some() 不会改变原始数组。
-            return Object.keys(data).some(key => {
-              // indexOf() 返回某个指定的字符在某个字符串中首次出现的位置，如果没有找到就返回-1；
-              // 该方法对大小写敏感！所以之前需要toLowerCase()方法将所有查询到内容变为小写。
-              return String(data[key]).indexOf(search) > -1
-            })
-          })
-        }
-        return this.datasInfos;
-      },
+    
     },
     methods: 
     {
@@ -593,24 +566,27 @@ export default
       },
       handleChangeInter(file, fileList) {
        // this.fileList = fileList.slice(-3);
+       this.fileList=fileList;
        console.log("fileList="+fileList);
        this.taskInfo.descFile.interfaceFile=fileList.slice(-3)[0].name;
-       fileList=[];
+       this.fileList=[];
        // this.taskInfo.descFile.interfaceFile=fileList.slice(-3)[0].name;
         //console.log("interfacename:"+this.taskInfo.descFile.interfaceFile);
       },
        handleChangeSec(file, fileList) {
+          this.fileList=fileList;
         //this.fileList = fileList.slice(-3);
          this.taskInfo.descFile.sectionFile=fileList.slice(-3)[0].name;
          //for(var i=0;i<fileList.slice(-3).length;i++) { this.taskInfo.descFile.sectionFile.push(fileList.slice(-3)[i].name);console.log(i+"="+fileList.slice(-3)[i].name); }
-          fileList=[];
+          this.fileList=[];
        
       },
        handleChangeReg(file, fileList) {
+          this.fileList=fileList;
     
         this.taskInfo.descFile.registerFile=fileList.slice(-3)[0].name;;
       
-          fileList=[];
+          this.fileList=[];
        
       },
       uploadOnError(){
@@ -752,7 +728,7 @@ export default
               }
              
              //清空文件列表
-              fileList= []; 
+               this.fileList= []; 
                 this.taskInfo = 
                 {
                   taskName: '',
